spritey.core/test/spritey/core/validator/CharacterValidatorTests.java
/**
 * This source file is part of Spritey - the sprite sheet creator.
 * 
 * Copyright 2010 Maksym Bykovskyy.
 * 
 * Spritey is free software: you can redistribute it and/or modify it under the
 * terms of the GNU General Public License as published by the Free Software
 * Foundation, either version 3 of the License, or (at your option) any later
 * version.
 * 
 * Spritey is distributed in the hope that it will be useful, but WITHOUT ANY
 * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
 * A PARTICULAR PURPOSE. See the GNU General Public License for more details.
 * 
 * You should have received a copy of the GNU General Public License along with
 * Spritey. If not, see <http://www.gnu.org/licenses/>.
 */
package spritey.core.validator;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertTrue;

import org.junit.Before;
import org.junit.Test;

/**
 * Tests the implementation of CharacterValidator.
 */
public class CharacterValidatorTests {

    static final String LEGAL_CHARACTERS = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789-_";

    Validator validator;

    @Before
    public void initialize() {
        validator = new CharacterValidator(LEGAL_CHARACTERS);
    }

    @Test
    public void isValidIllegalCharacters() {
        assertFalse(validator.isValid("add.png"));
        assertEquals(CharacterValidator.ILLEGAL_CHARACTER,
                validator.getErrorCode());
        assertNotNull(validator.getMessage());
        assertFalse(validator.getMessage().isEmpty());

        assertFalse(validator.isValid("button_pressed.png"));
        assertEquals(CharacterValidator.ILLEGAL_CHARACTER,
                validator.getErrorCode());
        assertNotNull(validator.getMessage());
        assertFalse(validator.getMessage().isEmpty());

        assertFalse(validator.isValid("sprite 1"));
        assertEquals(CharacterValidator.ILLEGAL_CHARACTER,
                validator.getErrorCode());
        assertNotNull(validator.getMessage());
        assertFalse(validator.getMessage().isEmpty());

        assertFalse(validator.isValid("New Group"));
        assertEquals(CharacterValidator.ILLEGAL_CHARACTER,
                validator.getErrorCode());
        assertNotNull(validator.getMessage());
        assertFalse(validator.getMessage().isEmpty());
    }

    @Test
    public void isValidLegalCharacters() {
        assertTrue(validator.isValid("add_png"));
        assertEquals(Validator.NONE, validator.getErrorCode());
        assertNotNull(validator.getMessage());
        assertTrue(validator.getMessage().isEmpty());

        assertTrue(validator.isValid("12345"));
        assertEquals(Validator.NONE, validator.getErrorCode());
        assertNotNull(validator.getMessage());
        assertTrue(validator.getMessage().isEmpty());

        assertTrue(validator.isValid("_sprite"));
        assertEquals(Validator.NONE, validator.getErrorCode());
        assertNotNull(validator.getMessage());
        assertTrue(validator.getMessage().isEmpty());

        assertTrue(validator.isValid("button-pressed"));
        assertEquals(Validator.NONE, validator.getErrorCode());
        assertNotNull(validator.getMessage());
        assertTrue(validator.getMessage().isEmpty());

        assertTrue(validator.isValid("sprite2"));
        assertEquals(Validator.NONE, validator.getErrorCode());
        assertNotNull(validator.getMessage());
        assertTrue(validator.getMessage().isEmpty());
    }

}
